Refactor authentication middleware and route handling. Update path extraction logic to improve locale handling and enhance route protection checks. Implement static generation for several pages to optimize performance.
