+++ app/[locale]/(Home)/policy/page.tsx
@@ -11,6 +11,9 @@ export const metadata: Metadata = {
   },
 }
 
+// Force static generation for this route
+export const dynamic = 'force-static'
+
 const Policy = async ({ params }: { params: Promise<{ locale: string }> }) => {
   const { locale } = await params
   return (
