+++ lib/appRoutes.ts
@@ -1,8 +1,25 @@
 export const AUTH_PATHS = ['/signUp', '/signIn']
 export const PUBLIC_PATHS = ['/']
-export const PRIVATE_PATHS = ['']
+export const PRIVATE_PATHS = [
+  '/profile',
+  '/posts/createNewPost',
+  '/posts/allPosts',
+  '/posts/editPost',
+  '/events/allEvents',
+  '/events/createEvent',
+  '/events/createEventCategory',
+  '/events/createEventSeries',
+  '/events/editEvent',
+  '/events/manageSponsors',
+]
 export const ADMIN_PATHS = [
   '/posts/createNewPost',
   '/posts/allPosts',
   '/posts/editPost',
+  '/events/allEvents',
+  '/events/createEvent',
+  '/events/createEventCategory',
+  '/events/createEventSeries',
+  '/events/editEvent',
+  '/events/manageSponsors',
 ]
